requirements.txt
aws-cdk-lib==2.110.1
constructs>=10.0.0,<11.0.0
boto3>=1.34.0
aws-lambda-powertools>=2.25.0

# Testing dependencies
pytest>=7.4.0
hypothesis>=6.88.0
moto>=4.2.0